Created FK/PK for subset tables TTDRacks/BDDRacks. Removed quantity fields, see full commit for details
Since each rack is referenced by a unique Equipment ID, the quantity fields for each size of rack will cause duplicate data. Quantity by size can be found through aggregate functions.
The extra data the subset tables hold is the size of the racks, which is specific to TTDRacks/BDDRacks and would not make sense to be included in the more general Equipment table.
Quantity field specs were moved to non-used specs folder, and NOT deleted.