Adjustments to SOInfo table. See full commit for details
Renamed SOInfos to SOSets, which is more clear and indicates that each record will be an SO Set.
Paul wanted to be able to not only reference SO Set sizes, but also be able to keep track of each set
within a size (They are labeled with Alphabet letters, in sequence of A, B, C, ... AA, AB, AC, ... etc). This is achieved by
adding an SOSetLabel spec.
So the SO Sets table now is a subset for equipment, meaning SOInfoID was changed to SOSetID and is a PK + FK (EquipmentID FK)
SO#OfSets within a size is removed, can be recreated by aggregate function that counts all sets of a particular size.